9.9 Función ejecutarOperacion y pruebas video 10
#134

+++ Tecnicatura3J/Lecion09/CalculadoraUTN/src/CalculadoraUTN.java
@@ -43,10 +43,10 @@ public class CalculadoraUTN {
         }//Fin metodo mostrar menu
     private static void ejecutarOperacion(int operacion, Scanner entrada){
         System.out.print("Digite el valor para operando1: ");
-        var operando1 = Integer.parseInt(entrada.nextLine());
+        var operando1 = Double.parseDouble(entrada.nextLine());
         System.out.print("Digite el valor para operando2: ");
-        var operando2 = Integer.parseInt(entrada.nextLine());
-        int resultado;
+        var operando2 = Double.parseDouble(entrada.nextLine());
+        Double resultado;
         switch (operacion) {
             case 1 -> { //Suma
                 resultado = operando1 + operando2;
